Check for section link in link fixing
Section links weren't working because the id's in the page are all lowercase, and are set to uppercase in the markup. This checks for section links by checking for a # as the first character in the string. If this is true, it converts the link target to lowercase, linking to the sections correctly. Then I pass this as a flag to the function making links absolute, to skip over making this link absolute if it has already been modified.