Allow subdirectories!
This will let us have other projects (SDL_mixer, SDL_image, etc) in their
own sections, which helps for automated tools and general organization.

You need to manually make the subdir at the moment, either by pushing
something through git or just mkdir'ing in the server's "raw" directory;
since subdir creation is currently intended to be an extremely rare event
that an admin does, this is acceptable for now, but a basic UI for this
would be nice eventually.