edit: Use sessionStorage to keep the current textarea contents.
This makes your in-progress work survive reloads, etc, as long as the
browser tab lives, so jumping to a different page and then hitting back,
or pressing the reload button, won't throw away your edits, so we don't
have to have a popup warning if this works.

Clicking "Cancel" or submitting an edit will dump the sessionStorage
item, so you don't end up with old/rejected text later in this case.

Also, this uses sessionStorage, so it will not save your work
long-term! This seems more sensible, since we're trying to save against
accidental loss in the moment and not saving drafts, and don't want
_really_ old work popping up and replacing newer revisions.