Edits that generate pull requests can now be added to.
Future edits will notice branches that still exist and push new changes to the
same PR, letting the editor either start from their existing changes or start
over with the latest page.

This means when the user goes "oh whoops, I made a typo," and clicks "edit"
again, it won't generate a totally new PR, they won't have to retype their
previous changes before making further correction.

This relies on merged/rejected PRs explicitly deleting the topic branch, which
is one extra click in the GitHub UI. Otherwise, future edits from that user
to that page will still append to the completed PR branch. This might be
improved in the future, if we can ask GitHub if a branch is finished in some
way.

Fixes #41.